Refactor and move files
