+++ SeQuant/core/optimize/optimize.cpp
@@ -1,4 +1,4 @@
-#include "optimize.hpp"
+#include "SeQuant/core/optimize.hpp"
 
 namespace sequant::optimize {
 
@@ -35,17 +35,19 @@ void pull_scalar(sequant::ExprPtr expr) noexcept {
   prod.scale(scal);
 }
 
-EvalNode optimize(const ExprPtr& expr, bool canonize) {
+EvalNode optimize(const ExprPtr& expr) {
   using ranges::views::transform;
   if (expr->is<Tensor>())
     return to_eval_node(expr);
   else if (expr->is<Product>()) {
+    // canonicalization within sto doesn't seem beneficial
+    bool canonize = false;
     return single_term_opt(expr->as<Product>(), canonize)
         .optimal_seqs.begin()
         ->clone();
   } else if (expr->is<Sum>()) {
-    auto smands = *expr | transform([canonize](auto const& s) {
-      return to_expr(optimize(s, canonize));
+    auto smands = *expr | transform([](auto const& s) {
+      return to_expr(optimize(s));
     }) | ranges::to_vector;
 
     return to_eval_node(ex<Sum>(Sum{smands.begin(), smands.end()}));
