+++ examples/eval/calc_info.cpp
@@ -9,7 +9,7 @@ namespace sequant::eval {
 
 EvalNode CalcInfo::node_(const ExprPtr& expr, size_t rank) const {
   auto trimmed = optimize::tail_factor(expr);
-  return optm_opts.single_term ? optimize::optimize(trimmed, false)
+  return optm_opts.single_term ? optimize::optimize(trimmed)
                                : to_eval_node(trimmed);
 }
